server/requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.24
Django==1.10.2
djangorestframework==3.5.0
kombu==3.0.37
paho-mqtt==1.2
pytz==2016.7
